chore: Refractored type `PrintMode`
